Factor ReportImageButton out of UserImage
